refactor: :recycle: Refactored web modules
